+++ cpp/src/toppra/solver/seidel.cpp
@@ -1,46 +1,15 @@
 #include <toppra/solver/seidel.hpp>
+#include <toppra/solver/seidel-internal.hpp>
 
 namespace toppra {
 namespace solver {
 
-typedef Eigen::Matrix<value_type, 2, 1> Vector2;
-typedef Eigen::Matrix<value_type, 1, 2> RowVector2;
-typedef Eigen::Matrix<value_type, 3, 1> Vector3;
-typedef Eigen::Matrix<value_type, 1, 3> RowVector3;
-typedef Eigen::Matrix<value_type, Eigen::Dynamic, 2> MatrixX2;
-typedef Eigen::Matrix<value_type, Eigen::Dynamic, 3> MatrixX3;
-
 namespace seidel {
-struct LpSol {
-  bool feasible;
-  value_type optval;
-  Vector2 optvar;
-  std::array<int, 2> active_c;
-};
-
-std::ostream& operator<< (std::ostream& os, const LpSol& sol)
-{
-  if (!sol.feasible) return os << "infeasible";
-  os << "feasible: " << sol.optval
-    << "\n\toptvar: " << sol.optvar.transpose()
-    //<< "\n\tactive_c: " << sol.active_c
-    ;
-}
-
 const LpSol INFEASIBLE { false };
 
-constexpr double TINY = 1e-10;
-constexpr double SMALL = 1e-8;
-
-// bounds on variable used in seidel solver wrapper. u and x at every
-// stage is constrained to stay within this range.
-constexpr double VAR_MIN = -1e8;
-constexpr double VAR_MAX =  1e8;
-
-// absolute largest value that a variable can have. This bound should
-// be never be reached, however, in order for the code to work properly.
-constexpr double INF = 1e10;
-
+/// Compute value coeffs * [v 1] of a constraint.
+/// handling infinite values for v as, when v[i] is +/- infinity,
+/// c[i]*v[i] == 0 if v[i] == 0.
 template<class Coeffs, class Vars>
 typename Coeffs::Scalar value(const Eigen::MatrixBase<Coeffs>& coeffs,
     const Eigen::MatrixBase<Vars>& vars)
@@ -49,122 +18,41 @@ typename Coeffs::Scalar value(const Eigen::MatrixBase<Coeffs>& coeffs,
       && Vars::ColsAtCompileTime == 1
       && Coeffs::ColsAtCompileTime == Vars::RowsAtCompileTime + 1,
       "Size mismatch between coefficient (1x(N+1)) and vars (Nx1)");
-  return coeffs.template head<Vars::RowsAtCompileTime>() * vars + coeffs[Coeffs::ColsAtCompileTime-1];
-}
-
-/**
-Solve a Linear Program with 1 variable.
-
-max   v[0] x + v[1]
-s.t.  A [ x 1 ] <= 0
-**/
-template<typename Derived>
-LpSol solve_lp1d(const RowVector2& v, const Eigen::MatrixBase<Derived>& A)
-{
-  value_type cur_min = -INF,
-             cur_max = INF;
-  int active_c_min = -1,
-      active_c_max = -2;
-  auto a (A.col(0)), b (A.col(1));
-
-  TOPPRA_LOG_DEBUG("Seidel LP 1D:\n"
-      << "v: " << v << '\n'
-      << "A:\n" << A << '\n'
-      );
-
-  for (int i = 0; i < a.size(); ++i) {
-    if (a[i] > TINY) {
-      value_type cur_x = - b[i] / a[i];
-      if (cur_x < cur_max) {
-        cur_max = cur_x;
-        active_c_max = i;
-      }
-    } else if (a[i] < -TINY) {
-      value_type cur_x = - b[i] / a[i];
-      if (cur_x > cur_min) {
-        cur_min = cur_x;
-        active_c_min = i;
-      }
-    } else if (b[i] > SMALL) {
-      TOPPRA_LOG_DEBUG("Seidel LP 1D: constraint " << i << " infeasible.");
-      return INFEASIBLE;
-    }
-    // else a[i] is approximately zero. do nothing.
-  }
-
-  if (cur_min - cur_max > SMALL)
-    return INFEASIBLE;
-
-  if (abs(v[0]) < TINY || v[0] < 0) {
-    // optimizing direction is perpencicular to the line, or is
-    // pointing toward the negative direction.
-    return LpSol{ true,
-      v[0] * cur_min + v[1],
-      { cur_min, 0. },
-      { active_c_min, false },
-    };
-  } else {
-    return LpSol{ true,
-      v[0] * cur_max + v[1],
-      { cur_max, 0. },
-      { active_c_max, 0 },
-    };
-  }
+  typename Coeffs::Scalar res = coeffs(coeffs.cols()-1);
+  for (int i = 0; i < Vars::RowsAtCompileTime; ++i)
+    res += (coeffs[i]==0 && !std::isfinite(vars[i]) ? 0 : coeffs[i]*vars[i]);
+  return res;
 }
 
 namespace internal {
   // projective coefficients to the line
   // add respective coefficients to A_1d
-  template<typename Derived>
-  void compute_denom_and_t_limit (const Eigen::MatrixBase<Derived>& Aj,
+  template<typename Derived, typename Derived2>
+  inline void project_linear_constraint (const Eigen::MatrixBase<Derived>& Aj,
       const Vector2& d_tan, const Vector2& zero_prj,
-      value_type& denom, value_type& t_limit)
+      const Eigen::MatrixBase<Derived2>& Aj_1d_)
   {
-    denom = Aj.template head<2>() * d_tan;
-    t_limit = value(Aj, zero_prj);
+    Derived2& Aj_1d = const_cast<Derived2&>(Aj_1d_.derived());
+    Aj_1d <<
+      Aj.template head<2>() * d_tan,
+      value(Aj, zero_prj);
   }
 }
 
-/**
-Solve a LP with two variables.
+#define TOPPRA_SEIDEL_LP2D(w,X)                         \
+  TOPPRA_LOG_##w("Seidel LP 2D:\n"                      \
+      << "v: " << v << '\n'                             \
+      << "A:\n" << A << '\n'                            \
+      << low[0] << "\t<= x[0] <= " << high[0] << '\n'    \
+      << low[1] << "\t<= x[1] <= " << high[1] << '\n'    \
+      << X);
 
-The LP is specified as follow:
-        max     v^T x
-        s.t.    A [ x^T 1 ]^T <= 0
-                low <= x <= high
-
-NOTE: A possible optimization for this function is pruning linear
-constraints that are clearly infeasible. This is not implemented
-because in my current code, the bottleneck is not in solving
-TOPP-RA but in setting up the parameters.
-
-Parameters
-----------
-\param v
-\param a
-\param b
-\param c
-\param low
-\param high
-\param active_c Contains (2) indicies of rows in a, b, c that are likely the
-    active constraints at the optimal solution.
-\param use_cache: bool
-\param index_map A view to a pre-allocated integer array, to map from
-    [1,...,nrows] to the considered entries. This array is created
-    to avoid the cost of initializing a new array.
-\param A_1d A view to an initialized array. This array is created to avoid
-    the cost of initializing a new array.
-
-\return A LpSol instance that contains the ouput. When LpSol.feasible is false,
-        other fields are not meaningful.
-**/
-LpSol solve_lp2d(const RowVector2& v,
-    const MatrixX3& A,
+LpSol solve_lp2d(const RowVector2& v, const MatrixX3& A,
     const Vector2& low, const Vector2& high,
-    std::array<int, 2> active_c, bool use_cache,
-    std::vector<int> index_map,
     MatrixX2& A_1d)
 {
+  assert(A_1d.rows() == A.rows()+4);
+
   // number of working set recomputation
   unsigned int nrows = A.rows();
   Vector2 cur_optvar;
@@ -176,63 +64,58 @@ LpSol solve_lp2d(const RowVector2& v,
   LpSol sol;
 
   // print all input to the algorithm
-  TOPPRA_LOG_DEBUG("Seidel LP 2D:\n"
-      << "v: " << v << '\n'
-      << "A:\n" << A << '\n'
-      << "lo: " << low .transpose() << '\n'
-      << "hi: " << high.transpose() << '\n'
-      );
-
-  if (use_cache) {
-    assert(index_map.size() == nrows);
-  } else {
-    index_map.resize(nrows);
-    for (int i = 0; i < nrows; ++i) index_map[i] = i;
-    A_1d = MatrixX2::Zero(nrows+4, 2);
-  }
+  TOPPRA_SEIDEL_LP2D(DEBUG, "");
 
   // handle fixed bounds (low, high). The following convention is
   // adhered to: fixed bounds are assigned the numbers: -1, -2, -3,
   // -4 according to the following order: low[0], high[0], low[1],
   // high[1].
-  if ((low.array() > high.array()).any()) {
-    TOPPRA_LOG_DEBUG("Seidel: incoherent bounds.");
-    return INFEASIBLE;
-  }
   for (int i = 0; i < 2; ++i) {
-    if (v[i] > TINY) {
-      cur_optvar[i] = high[i];
-      sol.active_c[i] = (i == 0) ? -2 : -4;
-    } else {
-      cur_optvar[i] = low[i];
-      sol.active_c[i] = (i == 0) ? -1 : -3;
+    if (low[i] > high[i]) {
+      // If the difference between low and high is sufficiently small, then
+      // return infeasible.
+      if (   low[i] - high[i] > std::max(std::abs(low[i]),std::abs(high[i]))* REL_TOLERANCE
+          || low[i] == infinity
+          || high[i] == -infinity) {
+        TOPPRA_SEIDEL_LP2D(WARN,
+            "-> incoherent bounds. high - low = " << (high - low).transpose());
+        return INFEASIBLE;
+      }
+      // Otherwise, assume variable i is static. Thus we are left with a 1D LP.
+      int j = 1-i;
+      sol.optvar[i] = (low[i]+high[i])/2;
+      A_1d.topRows(nrows) << A.col(j), A.col(2) + sol.optvar[i] * A.col(i);
+      A_1d.middleRows<2>(nrows) << -1,   low[j],
+                                    1, -high[j];
+
+      LpSol1d sol_1d = solve_lp1d({ v[j], 0. }, A_1d.topRows(nrows+2));
+      if (!sol_1d.feasible) {
+        TOPPRA_SEIDEL_LP2D(WARN, "-> infeasible");
+        return INFEASIBLE;
+      }
+      sol.optvar[j] = sol_1d.optvar;
+      sol.feasible = true;
+      sol.optval = v * sol.optvar;
+      sol.active_c[0] = v[i] > 0 ? HIGH(i) : LOW(i);
+      switch(sol_1d.active_c - nrows) {
+        case 0: sol.active_c[1] = LOW (j); break;
+        case 1: sol.active_c[1] = HIGH(j); break;
+        default: sol.active_c[1] = sol_1d.active_c;
+      }
+      return sol;
     }
+
+    cur_optvar[i]   = v[i] > 0 ? high[i] : low[i];
+    sol.active_c[i] = v[i] > 0 ? HIGH(i) : LOW(i);
   }
   TOPPRA_LOG_DEBUG("cur_optvar = " << cur_optvar.transpose());
-
-  // If active_c contains valid entries, swap the first two indices
-  // in index_map to these values.
-  unsigned int cur_row = 2;
-  if (   active_c[0] >= 0 && active_c[0] < nrows
-      && active_c[1] >= 0 && active_c[1] < nrows
-      && active_c[0] != active_c[1]) {
-    // active_c contains valid indices
-    index_map[0] = active_c[1];
-    index_map[1] = active_c[0];
-    for (int i = 0; i < nrows; ++i)
-      if (i != active_c[0] && i != active_c[1])
-        index_map[cur_row++] = i;
-  } else
-    for (int i = 0; i < nrows; ++i)
-      index_map[i] = i;
 
   // pre-process the inequalities, remove those that are redundant
 
   // handle other constraints in a, b, c
-  for (int k = 0; k < nrows; ++k) {
-    int i = index_map[k];
+  for (int i = 0; i < nrows; ++i) {
     // if current optimal variable satisfies the i-th constraint, continue
-    if (value(A.row(i), cur_optvar) < TINY)
+    if (value(A.row(i), cur_optvar) < ABS_TOLERANCE)
       continue;
     // otherwise, project all constraints on the line defined by (a[i], b[i], c[i])
     sol.active_c[0] = i;
@@ -250,80 +133,53 @@ LpSol solve_lp2d(const RowVector2& v,
     Vector2 d_tan { -A(i,1), A(i,0) }; // vector parallel to the line
     Vector2 v_1d { v * d_tan, 0 };
 
-    // project 4 + k constraints onto the parallel line. each
-    // constraint occupies a row of A_1d.
-    for (int j = 0; j < 4 + k; ++j) { // nb rows 1d.
-      value_type denom, t_limit;
-      switch (j-k) {
-        case 0: // j == k: handle low <= x
-          internal::compute_denom_and_t_limit(RowVector3 { -1., 0., low[0] },
-              d_tan, zero_prj, denom, t_limit);
-          break;
-        case 1: // j == k + 1: handle x <= high
-          internal::compute_denom_and_t_limit(RowVector3 { 1., 0., -high[0] },
-              d_tan, zero_prj, denom, t_limit);
-          break;
-        case 2: // j == k + 2: handle low <= y
-          internal::compute_denom_and_t_limit(RowVector3 { 0., -1., low[1] },
-              d_tan, zero_prj, denom, t_limit);
-          break;
-        case 3: // j == k + 3: handle y <= high
-          internal::compute_denom_and_t_limit(RowVector3 { 0., 1., -high[1] },
-              d_tan, zero_prj, denom, t_limit);
-          break;
-        default: // handle other constraint
-          internal::compute_denom_and_t_limit(A.row(index_map[j]),
-              d_tan, zero_prj, denom, t_limit);
-      }
-
-      if (denom > TINY)
-        A_1d.row(j) <<  1.,  t_limit / denom;
-      else if (denom < -TINY)
-        A_1d.row(j) << -1., -t_limit / denom;
-      else {
-        // Current constraint is parallel to the base one.
-        // if infeasible, return failure.
-        if (t_limit > SMALL) {
-          TOPPRA_LOG_DEBUG("Seidel: infeasible constraint. t_limit = " << t_limit
-              << ", denom = " << denom);
-          return INFEASIBLE;
-        }
-        // Otherwise, specify 0 <= 1
-        A_1d.row(j) << 0, -1.;
-      }
-    }
+    // Size of the 1D LP: 4 + k
+    // Compute the constraint parameters of the 1D LP corresponding to the
+    // linear constraints.
+    for (int j = 0; j < i; ++j)
+      internal::project_linear_constraint(A.row(j),
+              d_tan, zero_prj, A_1d.row(j));
+    //A_1d.topRows(k) << A.topRows(k) * d_tan, value(A.topRows(k), zero_prj);
+    // Compute the constraint parameters of the 1D LP corresponding to the
+    // 4 bound constraints.
+    // handle low <= x
+    internal::project_linear_constraint(RowVector3 { -1., 0., low[0] },
+        d_tan, zero_prj, A_1d.row(i));
+    // handle x <= high
+    internal::project_linear_constraint(RowVector3 { 1., 0., -high[0] },
+        d_tan, zero_prj, A_1d.row(i+1));
+    // handle low <= y
+    internal::project_linear_constraint(RowVector3 { 0., -1., low[1] },
+        d_tan, zero_prj, A_1d.row(i+2));
+    // handle y <= high
+    internal::project_linear_constraint(RowVector3 { 0., 1., -high[1] },
+        d_tan, zero_prj, A_1d.row(i+3));
 
     // solve the projected, 1 dimensional LP
     TOPPRA_LOG_DEBUG("Seidel LP 2D activate constraint " << i);
-    LpSol sol_1d = solve_lp1d(v_1d, A_1d.topRows(4+k));
+    LpSol1d sol_1d = solve_lp1d(v_1d, A_1d.topRows(4+i));
     TOPPRA_LOG_DEBUG("Seidel LP 1D solution:\n" << sol_1d);
 
-    // 1d lp infeasible
-    if (!sol_1d.feasible) return INFEASIBLE;
+    if (!sol_1d.feasible) {
+      TOPPRA_SEIDEL_LP2D(WARN, "-> infeasible");
+      return INFEASIBLE;
+    }
 
     // feasible, wrapping up
     // compute the current optimal solution
-    cur_optvar = zero_prj + sol_1d.optvar[0] * d_tan;
-    TOPPRA_LOG_DEBUG("cur_optvar = " << cur_optvar.transpose());
+    cur_optvar = zero_prj + sol_1d.optvar * d_tan;
+    TOPPRA_LOG_DEBUG("i = " << i << ". cur_optvar = " << cur_optvar.transpose());
     // record the active constraint's index
-    switch (sol_1d.active_c[0] - k) {
-      case 0: sol.active_c[1] = -1; break;
-      case 1: sol.active_c[1] = -2; break;
-      case 2: sol.active_c[1] = -3; break;
-      case 3: sol.active_c[1] = -4; break;
-      default:
-              if (sol_1d.active_c[0] < k)
-                sol.active_c[1] = index_map[sol_1d.active_c[0]];
-              else
-                // the algorithm should not reach this point. If it
-                // does, this means the active constraint at the
-                // optimal solution is the fixed bound used in the 1
-                // dimensional subproblem. This should not happen
-                // though.
-                return INFEASIBLE;
-    }
+    assert(sol_1d.active_c >= -2 && sol_1d.active_c < i+4);
+    if (sol_1d.active_c < 0) // Unbounded
+      sol.active_c[i] = sol_1d.active_c;
+    else if (sol_1d.active_c >= i) // Bound constraint
+      sol.active_c[1] = i - sol_1d.active_c - 1;
+    else // Linear constraint
+      sol.active_c[1] = sol_1d.active_c;
   }
 
+  // Sanity check for debugging purpose.
   for (int i = 0; i < nrows; ++i) {
     value_type v = value(A.row(i), cur_optvar);
     if (v > 0) {
@@ -340,6 +196,8 @@ LpSol solve_lp2d(const RowVector2& v,
   sol.optval = v * sol.optvar;
   return sol;
 }
+
+#undef TOPPRA_SEIDEL_LP2D
 
 }
 
@@ -362,8 +220,9 @@ void Seidel::initialize (const LinearConstraintPtrs& constraints, const Geometri
   // m_A, m_low, m_high.
   // The first dimensions of these array all equal N + 1.
   m_A.assign(N+1, MatrixX3::Zero(nC, 3));
-  m_low  = MatrixX2::Constant(N + 1, 2, seidel::VAR_MIN);
-  m_high = MatrixX2::Constant(N + 1, 2, seidel::VAR_MAX);
+  m_A_ordered.resize(nC, 3);
+  m_low  = MatrixX2::Constant(N + 1, 2, -seidel::infinity);
+  m_high = MatrixX2::Constant(N + 1, 2,  seidel::infinity);
   int cur_index = 2;
 
   for (const Solver::LinearConstraintParams& p : m_constraintsParams.lin) {
@@ -399,8 +258,8 @@ void Seidel::initialize (const LinearConstraintPtrs& constraints, const Geometri
   // init constraint coefficients for the 1d LPs
   m_A_1d = MatrixX2::Zero(nC + 4, 2);
   m_index_map.resize(nC, 0);
-  m_active_c_up.fill(0);
-  m_active_c_down.fill(0);
+  m_active_c_up.fill(-1);
+  m_active_c_down.fill(-1);
 }
 
 bool Seidel::solveStagewiseOptim(std::size_t i,
@@ -425,11 +284,11 @@ bool Seidel::solveStagewiseOptim(std::size_t i,
   // handle x_next_min <= 2 delta u + x_i <= x_next_max
   if (i < N) {
     value_type delta = deltas()[i];
-    if (xNext[0] < -seidel::INF) // TODO isnan(x_next_min)
+    if (xNext[0] <= -seidel::infinity) // TODO isnan(x_next_min)
       m_A[i].row(0) << 0, 0, -1;
     else
       m_A[i].row(0) << -2*delta, -1, xNext[0];
-    if (xNext[1] > seidel::INF) // TODO isnan(x_next_max)
+    if (xNext[1] >= seidel::infinity) // TODO isnan(x_next_max)
       m_A[i].row(1) << 0, 0, -1;
     else
       m_A[i].row(1) << 2*delta, 1, -xNext[1];
@@ -446,15 +305,41 @@ bool Seidel::solveStagewiseOptim(std::size_t i,
   // is be selected depending on the sign of g[1]
   bool upper (g[1] > 0);
   auto& active_c = (upper ? m_active_c_up : m_active_c_down);
+
+  // If active_c contains valid entries, swap the first two indices
+  // in index_map to these values.
+  if (active_c[0] != -1) {
+    assert(active_c[0] >= 0 && active_c[0] < m_A[i].rows()
+        && active_c[1] >= 0 && active_c[1] < m_A[i].rows()
+        && active_c[0] != active_c[1]);
+    // active_c contains valid indices
+    m_index_map[0] = active_c[1];
+    m_index_map[1] = active_c[0];
+    m_A_ordered.topRows<2>() << m_A[i].row(active_c[1]),
+                                m_A[i].row(active_c[0]);
+    for (int k = 0, cur_row = 2; k < m_A[i].rows(); ++k)
+      if (k != active_c[0] && k != active_c[1]) {
+        m_index_map[cur_row] = k;
+        m_A_ordered.row(cur_row++) = m_A[i].row(k);
+      }
+  } else {
+    for (int i = 0; i < m_A[i].rows(); ++i)
+      m_index_map[i] = i;
+    m_A_ordered = m_A[i];
+  }
+
   // solver selected:
   // - upper: when computing the lower bound of the controllable set.
   // - lower: when computing the lower bound of the controllable set,
   //          or computing the parametrization in the forward pass.
-  seidel::LpSol lpsol = seidel::solve_lp2d(v, m_A[i],
-      low, high, active_c, true, m_index_map, m_A_1d);
+  seidel::LpSol lpsol = seidel::solve_lp2d(v, m_A[i], low, high, m_A_1d);
   if (lpsol.feasible) {
     solution = lpsol.optvar;
-    active_c = lpsol.active_c;
+    if (lpsol.active_c[0] < 0 || lpsol.active_c[1] < 0)
+      active_c = {-1, -1};
+    else
+      for (int k = 0; k < 2; ++k)
+        active_c[k] = m_index_map[lpsol.active_c[k]];
     return true;
   }
   TOPPRA_LOG_DEBUG("Seidel: solver fails (upper ? " << upper << ')');