+++ cpp/src/toppra/toppra.hpp
@@ -17,6 +17,12 @@
 #define TOPPRA_LOG_DEBUG(X) ((void)0)
 #endif
 
+#if defined(TOPPRA_DEBUG_ON) || defined(TOPPRA_WARN_ON)
+#define TOPPRA_LOG_WARN(X) std::cout << "[WARN]: " <<  X << std::endl
+#else
+#define TOPPRA_LOG_WARN(X) ((void)0)
+#endif
+
 // Use for checking if a quantity is very close to zero
 #ifndef TOPPRA_NEARLY_ZERO
 #define TOPPRA_NEARLY_ZERO 1e-8